Add C interface for LDPC and Cython code for LDPC/simtools
LDPC C interface does not offer a function to create a structure
decoder_t because it is a polimorphic type and it cannot be converted
to C structure. It only offers a function to decode a LLR vector, but
every time it is called, it must create the decoder object anew.